プレイヤーがスライムに攻撃するとスライムは無敵状態になり一定時間経過すると無敵状態が解除され4回攻撃がヒットするとスライムが消えるようになりました。
