+++ src/main/java/entity/Entity.java
@@ -401,7 +401,7 @@ public class Entity {
             }
 
             g2.drawImage(image, screenX, screenY, FrameApp.getTileSize(), FrameApp.getTileSize(), null);
-            
+
             g2.setComposite(AlphaComposite.getInstance(AlphaComposite.SRC_OVER, 1f));
         }
     }
